@@ -1,6 +1,5 @@
 package com.wikia.tibia;
 
-import com.wikia.tibia.objects.TibiaWikiBot;
 import com.wikia.tibia.usecases.FixCreatures;
 import com.wikia.tibia.usecases.FixItems;
 
@@ -12,16 +11,12 @@ public class Main {
 
     public static void main(String[] args) {
 
-        // initialize bot and login
-        TibiaWikiBot tibiaWikiBot = new TibiaWikiBot();
-        tibiaWikiBot.login();
-
         // first usecase
-        FixCreatures fixCreatures = new FixCreatures(tibiaWikiBot);
+        FixCreatures fixCreatures = new FixCreatures();
         fixCreatures.checkCreatures();
 
         // second usecase
-        FixItems fixItems = new FixItems(tibiaWikiBot);
+        FixItems fixItems = new FixItems();
         fixItems.checkItems();
     }
 }
