@@ -2,22 +2,24 @@ package com.wikia.tibia.usecases;
 
 import com.wikia.tibia.objects.Creature;
 import com.wikia.tibia.objects.Item;
-import com.wikia.tibia.objects.LootItem;
-import com.wikia.tibia.objects.WikiObject;
-import com.wikia.tibia.repositories.WikiArticleRepository;
-import net.sourceforge.jwbf.core.contentRep.Article;
-import net.sourceforge.jwbf.mediawiki.actions.queries.CategoryMembersSimple;
-import net.sourceforge.jwbf.mediawiki.bots.MediaWikiBot;
+import com.wikia.tibia.repositories.CreatureRepository;
+import com.wikia.tibia.repositories.ItemRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class FixCreatures {
 
-    private static final Logger log = LoggerFactory.getLogger(FixCreatures.class);
+    private static final Logger LOG = LoggerFactory.getLogger(FixCreatures.class);
 
     private static final List ITEMS_WITH_NO_DROPPEDBY_LIST = Arrays.asList("Gold Coin", "Platinum Coin");
     private static final List EVENT_ITEMS = Arrays.asList(
@@ -28,30 +30,71 @@ public class FixCreatures {
             "Party Hat",
             "Silver Raid Token",
             "Golden Raid Token"
-            );
+    );
     private static final List NON_EVENT_CREATURES_DROPPING_SNOWBALLS = Arrays.asList("Yeti", "Grynch Clan Goblin");
-    private static final String REGEX_DROPPED_BY = "\\{\\{Dropped By\\|(.*?)}}";
-    private static final boolean DEBUG_MODE = true;
+    private static final boolean DEBUG_MODE = false;
 
-    private WikiArticleRepository repository;
-    private Map<String, Item> itemPagesToUpdate = new HashMap<>();
+    private CreatureRepository creatureRepository;
+    private ItemRepository itemRepository;
+    private Map<String, Item> itemPagesToUpdate = new ConcurrentHashMap<>(); // item name, actual item (for easy lookup)
+    private List<Item> items = new ArrayList<>();
+    private List<Creature> creatures = new ArrayList<>();
 
-    public FixCreatures(MediaWikiBot mediaWikiBot) {
-        this.repository = new WikiArticleRepository(mediaWikiBot);
+    public FixCreatures() {
+        this.creatureRepository = new CreatureRepository();
+        this.itemRepository = new ItemRepository();
     }
 
+    /**
+     * 1. Get a list of Creatures and sort it.
+     * 2. Filter out deprecated or event creatures.
+     * 3. Filter out creatures with no loot listed.
+     * 4. For every creature, go over all its loot items.
+     * 5. For every loot item, add a backreference to the creature, unless already present.
+     * 6. Save all the changes.
+     * <p>
+     * The end result is a list of Items with their droppedby list extended.
+     */
     public void checkCreatures() {
-        CategoryMembersSimple categoryMembers = repository.getMembersFromCategory("Creatures");
+        getCreatures().stream()
+                .sorted(Comparator.comparing(Creature::getName))
+                .filter(Creature::notDeprecatedOrEvent)
+                .filter(creature -> creature.getLoot() != null && !creature.getLoot().isEmpty())
+                .peek(c -> LOG.debug("Processing creature: {}", c.getName()))
+                .forEach(creature -> creature.getLoot().stream()
+                        .map(lootItem -> {
+                            final Optional<Item> item = getItem(lootItem.getItemName());
+                            if (item.isEmpty()) {
+                                LOG.error("Could not find loot item with name '{}' from creature '{}' in collection" +
+                                        " of items.", lootItem.getItemName(), creature.getName());
+                            }
+                            return item.orElse(null);
+                        })
+                        .filter(Objects::nonNull)
+                        .forEach(item -> addCreatureToDroppedByListOfItem(creature, item))
+                );
 
-        for (String creaturePageName : categoryMembers) {
-            Creature creature = (Creature) repository.getWikiObject(creaturePageName);
-            if (creature != null && !creature.isDeprecatedOrEvent() && !creature.getLoot().isEmpty()) {
-                for (LootItem lootItem : creature.getLoot()) {
-                    checkIfCreatureNameIsPresent(creaturePageName, lootItem.getItemName());
-                }
-            }
-        }
         saveItemArticles();
+    }
+
+    private List<Creature> getCreatures() {
+        if (creatures == null || creatures.isEmpty()) {
+            creatures = creatureRepository.getCreatures();
+        }
+        return creatures;
+    }
+
+    private List<Item> getItems() {
+        if (items == null || items.isEmpty()) {
+            items = itemRepository.getItems();
+        }
+        return items;
+    }
+
+    private Optional<Item> getItem(String itemName) {
+        return getItems().stream()
+                .filter(i -> Objects.equals(i.getName(), itemName))
+                .findAny();
     }
 
     private boolean itemShouldBeAdded(String creaturePageName, String lootItemNamePrecise) {
@@ -67,58 +110,44 @@ public class FixCreatures {
         return true;
     }
 
-    private void checkIfCreatureNameIsPresent(String creaturePageName, String lootItem) {
-        Item item = null;
+    /**
+     * If the creature is not already on the droppedby list of the item (compare ignoring case!) and the item is
+     * eligible for adding, add it. Also sort it.
+     */
+    private void addCreatureToDroppedByListOfItem(Creature creature, Item item) {
+        if (item.getDroppedby().contains(creature.getName()) && itemShouldBeAdded(creature.getName(), item.getName())) {
+            LOG.info("Adding creature '{}' to droppedby list of item '{}'.", creature.getName(), item.getName());
 
-        if (itemPagesToUpdate.containsKey(lootItem)) {
-            item = itemPagesToUpdate.get(lootItem);
-        } else {
-            WikiObject wikiObject = repository.getWikiObject(lootItem);
-            if (wikiObject instanceof Item) {
-                item = (Item) wikiObject;
+            if (!itemPagesToUpdate.keySet().contains(item.getName())) {
+                // item not already in itemPages cache, add it
+                item.getDroppedby().add(creature.getName());
+                Collections.sort(item.getDroppedby());
+                itemPagesToUpdate.put(item.getName(), item);
+            } else {
+                // item already presetn in itemPages cache, get the existing item and add the new creature name
+                final Item existingItem = itemPagesToUpdate.get(item.getName());
+                existingItem.getDroppedby().add(creature.getName());
+                Collections.sort(existingItem.getDroppedby());
             }
-        }
-
-        if (item == null) {
-            return;
-        }
-
-        List<String> creatureNames = item.getDroppedby();
-
-        if (!creatureNames.contains(creaturePageName)) {
-            creatureNames.add(creaturePageName);
-            item.setDroppedby(creatureNames);
-            itemPagesToUpdate.put(item.getName(), item);
-            log.info("[bot] adding creature '{}' to item '{}'.", creaturePageName, item.getName());
-        }
-    }
-
-    private String updateCreatureToDroppedByList(List<String> creatureNames, String newCreature) {
-        creatureNames.add(newCreature);
-        Collections.sort(creatureNames);
-        String newDroppedByList = String.join("|", creatureNames);
-        return "{{Dropped By|" + newDroppedByList + "}}";
-    }
-
-    private void addMissingCreatureNameToDroppedByList(String creaturePageName, Article itemPage, String textToInsert) {
-        String itemPageText = itemPage.getText();
-        Pattern p = Pattern.compile(REGEX_DROPPED_BY);
-        Matcher m = p.matcher(itemPageText);
-        if (m.find()) {
-            String newArticleText = m.replaceAll(textToInsert);
-            itemPage.setText(newArticleText);
-            itemPage.setEditSummary(String.format("[bot] adding creature '%s' to item '%s'.", creaturePageName, itemPage.getTitle()));
-            //itemPagesToUpdate.put(itemPage.getTitle(), itemPage);
-            log.info("[bot] adding creature '{}' to item '{}'.", creaturePageName, itemPage.getTitle());
         }
     }
 
     private void saveItemArticles() {
-        log.info("If debug mode is disabled I am going to update {} item articles with missing creatures.", itemPagesToUpdate.size());
+        LOG.info("If debug mode is disabled, {} item articles are being edited NOW.", itemPagesToUpdate.size());
         if (!DEBUG_MODE) {
-            for (Map.Entry<String, Item> itemPage : itemPagesToUpdate.entrySet()) {
-                //repository.saveArticle(itemPage.getValue()); // TODO fix the saving process, wikiObject -> json -> Article
-            }
+            itemPagesToUpdate.forEach((key, value) -> {
+                itemRepository.saveItem(value, "[bot] adding missing creature(s) to droppedby list.");
+                        pauseForABit();
+                    }
+            );
+        }
+    }
+
+    private void pauseForABit() {
+        try {
+            Thread.sleep(200);
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
         }
     }
 }
