@@ -1,141 +1,143 @@
 package com.wikia.tibia.usecases;
 
-import com.wikia.tibia.repositories.WikiArticleRepository;
-import net.sourceforge.jwbf.core.contentRep.Article;
-import net.sourceforge.jwbf.mediawiki.actions.queries.CategoryMembersSimple;
-import net.sourceforge.jwbf.mediawiki.bots.MediaWikiBot;
+import com.wikia.tibia.objects.Creature;
+import com.wikia.tibia.objects.Item;
+import com.wikia.tibia.objects.LootItem;
+import com.wikia.tibia.repositories.CreatureRepository;
+import com.wikia.tibia.repositories.ItemRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class FixItems {
 
-    private static final Logger log = LoggerFactory.getLogger(FixItems.class);
+    private static final Logger LOG = LoggerFactory.getLogger(FixItems.class);
+    private static final boolean DEBUG_MODE = false;
+    private static final List<String> ORC_RAID_ITEMS = Arrays.asList("Amazon Armor", "Amazon Helmet", "Amazon Shield");
 
-    private static final String REGEX_DROPPED_BY = "\\{\\{Dropped By\\|(.*?)}}";
-    private static final String REGEX_FILLED_DROPPED_BY = "\\{\\{Dropped By\\|(.+?)}}";
-    private static final String REGEX_LOOT_ITEM_NAME = "\\{\\{Loot Item\\|(([^|}]+)(\\|([^|}]+))?(\\|([^|}]+))?)}}";
+    private CreatureRepository creatureRepository;
+    private ItemRepository itemRepository;
+    private List<Item> items = new ArrayList<>();
+    private List<Creature> creatures = new ArrayList<>();
+    private Map<String, Creature> creaturePagesToUpdate = new ConcurrentHashMap<>(); // creature name, actual creature
 
-    private static final String REGEX_DEPRECATED_OR_EVENT = "\\{\\{(Deprecated|Event)}}";
-    private static final boolean DEBUG_MODE = true;
-
-    private MediaWikiBot mediaWikiBot;
-    private WikiArticleRepository repository;
-    private Map<String, Article> creaturePagesToUpdate = new HashMap<>();
-
-    public FixItems(MediaWikiBot mediaWikiBot) {
-        this.mediaWikiBot = mediaWikiBot;
-        this.repository = new WikiArticleRepository(mediaWikiBot);
+    public FixItems() {
+        this.creatureRepository = new CreatureRepository();
+        this.itemRepository = new ItemRepository();
     }
 
+    /**
+     * 1. Get a list of Items and sort it.
+     * 2. Filter out deprecated or event items.
+     * 3. Filter out items with no dropped by listed.
+     * 4. For every item, go over all its droppedby creatures.
+     * 5. For every creature, add a backreference to the lootItem, unless already present.
+     * 6. Save all the changes.
+     * <p>
+     * The end result is a list of Creatures with their LootTable extended.
+     */
     public void checkItems() {
-        CategoryMembersSimple items = repository.getMembersFromCategory("Items");
 
-        for (String itemPageName : items) {
-            Article itemPage = repository.getArticle(itemPageName);
-            String articleText = itemPage.getText();
-            if (!articleDeprecatedOrEvent(articleText) && articleTextContainsFilledDroppedBy(articleText)) {
-                List<String> listOfDroppedByCreatures = makeListOfDroppedByCreatures(articleText);
-                for (String droppedByCreature : listOfDroppedByCreatures) {
-                    boolean itemNameIsPresent = checkIfItemNameIsPresent(itemPageName, droppedByCreature);
-                    if (itemNameIsPresent) {
-//                        String lootItems = extractLootItems(droppedByCreature);
-                        List<String> listOfLootItems = makeListOfLootItems(droppedByCreature, articleText);
-//                        String newLootItemList = updateItemToLootItemList(lootItems, itemPageName);
-//                        addMissingLootItemToLootItemList(itemPageName, droppedByCreature, newLootItemList);
-                    }
-                }
-            }
-        }
+        getItems().stream()
+                .sorted(Comparator.comparing(Item::getName))
+                .filter(Item::notDeprecatedOrEvent)
+                .filter(item -> item.getDroppedby() != null && !item.getDroppedby().isEmpty())
+                .peek(i -> LOG.debug("Processing item: {}", i.getName()))
+                .forEach(item -> item.getDroppedby().stream()
+                        .map(creatureName -> {
+                            final Optional<Creature> creature = getCreature(creatureName);
+                            if (creature.isEmpty()) {
+                                LOG.error("Could not find creature with name '{}' from item '{}' in collection" +
+                                        " of creatures.", creatureName, item.getName());
+                            }
+                            return creature.orElse(null);
+                        })
+                        .filter(Objects::nonNull)
+                        .forEach(creature -> addItemToLootTableOfCreature(item, creature))
+                );
+
         saveCreatureArticles();
     }
 
-    private boolean articleDeprecatedOrEvent(String articleText) {
-        Pattern p = Pattern.compile(REGEX_DEPRECATED_OR_EVENT);
-        Matcher m = p.matcher(articleText);
-        return m.find();
-    }
-
-    private boolean articleTextContainsFilledDroppedBy(String articleText) {
-        Pattern p = Pattern.compile(REGEX_FILLED_DROPPED_BY);
-        Matcher m = p.matcher(articleText);
-        return m.find();
-    }
-
-    private List<String> makeListOfDroppedByCreatures(String creaturePageText) {
-        Pattern p = Pattern.compile(REGEX_DROPPED_BY);
-        Matcher m = p.matcher(creaturePageText);
-        if (m.find()) {
-            return new ArrayList<>(Arrays.asList(m.group(1).split("\\|")));
+    private List<Creature> getCreatures() {
+        if (creatures == null || creatures.isEmpty()) {
+            creatures = creatureRepository.getCreatures();
         }
-        return new ArrayList<>();
+        return creatures;
     }
 
-    private String extractLootItems(String droppedByCreature) {
-        Article creaturePage = getCreaturePage(droppedByCreature);
-        String creaturePageText = creaturePage.getText();
-        Pattern p = Pattern.compile(REGEX_LOOT_ITEM_NAME);
-        Matcher m = p.matcher(creaturePageText);
-        if (m.find()) {
-            return m.group(0);
-        }
-        log.warn("No LootItem template encountered on creature '{}'.", droppedByCreature);
-        return null;
+    private Optional<Creature> getCreature(String creatureName) {
+        return getCreatures().stream()
+                .filter(i -> Objects.equals(i.getName(), creatureName))
+                .findAny();
     }
 
-    private List<String> makeListOfLootItems(String creaturePageName, String articleText) {
-        List<String> lootItems = new ArrayList<>();
-        String lootItemName;
-        Pattern pattern = Pattern.compile(REGEX_LOOT_ITEM_NAME);
-        Matcher matcher = pattern.matcher(articleText);
-        while (matcher.find()) {
-            if (matcher.group(6) != null) { // Match |{{Loot Item|0-4|Gold Coin|very rare}}
-                lootItemName = matcher.group(4);
-            } else if (matcher.group(3) == null) { // Match |{{Loot Item|Apple}}
-                lootItemName = matcher.group(2);
-            } else if (Character.isUpperCase(matcher.group(2).charAt(0))) { // Match |{{Loot Item|Apple|always}}
-                lootItemName = matcher.group(2);
-            } else {  // Match |{{Loot Item|0-4|Gold Coin}}
-                lootItemName = matcher.group(4);
-            }
-                lootItems.add(lootItemName);
+
+    private List<Item> getItems() {
+        if (items == null || items.isEmpty()) {
+            items = itemRepository.getItems();
         }
-        return lootItems;
+        return items;
     }
 
     /**
-     * Get the Wiki Article from a specific creature, either the already updated cached version, or a fresh copy from the internet
+     * If the item is not already in the loot table of the creature and the creature is eligible for adding, add it.
      */
-    Article getCreaturePage(String droppedByCreature) {
-        Article creaturePage;
-        if (creaturePagesToUpdate.containsKey(droppedByCreature)) {
-            creaturePage = creaturePagesToUpdate.get(droppedByCreature);
-        } else {
-            creaturePage = repository.getArticle(droppedByCreature);
+    private void addItemToLootTableOfCreature(Item item, Creature creature) {
+        if (!creature.getLoot().contains(LootItem.fromName(item.getName())) && itemShouldBeAdded(creature.getName(), item.getName())) {
+            LOG.info("Adding item '{}' to loot table of creature '{}'.", item.getName(), creature.getName());
+
+            if (!creaturePagesToUpdate.keySet().contains(creature.getName())) {
+                // creature not already in creaturePages cache, add it
+                creature.getLoot().add(LootItem.fromName(item.getName()));
+                creaturePagesToUpdate.put(creature.getName(), creature);
+            } else {
+                // creature already present in creaturePages cache, get the existing creature and add the new lootItem
+                final Creature existingCreature = creaturePagesToUpdate.get(creature.getName());
+                existingCreature.getLoot().add(LootItem.fromName(item.getName()));
+            }
         }
-        return creaturePage;
+    }
+
+    /**
+     * Exclude certain items/creatures.
+     */
+    private boolean itemShouldBeAdded(String creaturePageName, String lootItemNamePrecise) {
+        if ("Orc Warlord".equals(creaturePageName)) {
+            return !ORC_RAID_ITEMS.contains(lootItemNamePrecise);
+        }
+
+        if ("Orc Tusk".equals(lootItemNamePrecise)) {
+            return !creaturePageName.contains("Orc");
+        }
+
+        return true;
     }
 
     private void saveCreatureArticles() {
-        // TODO implement this method
+        LOG.info("If debug mode is disabled, {} creature articles are being edited NOW.", creaturePagesToUpdate.size());
+        if (!DEBUG_MODE) {
+            creaturePagesToUpdate.forEach((key, value) -> {
+                        creatureRepository.saveCreature(value, "[bot] adding missing item(s) to loot list.");
+                        pauseForABit();
+                    }
+            );
+        }
     }
 
-    /**
-     * Check if item (e.g. Abyss Hammer) is present in the Loot Item list of given creature (e.g. Ferumbras)
-     */
-    private boolean checkIfItemNameIsPresent(String itemPageName, String droppedByCreature) {
-        String lootItems = extractLootItems(droppedByCreature);
-        return lootItems != null && !lootItems.contains(itemPageName);
-    }
-
-    private void addMissingLootItemToLootItemList(String itemPageName, String creaturePageName, String newLootItemList) {
-        Article creaturePage = getCreaturePage(creaturePageName);
-    }
-
-    private String updateItemToLootItemList(String lootItems, String itemPageName) {
-        return null;
+    private void pauseForABit() {
+        try {
+            Thread.sleep(200);
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
     }
 }
