@@ -1,22 +1,20 @@
 package com.wikia.tibia.usecases;
 
-import com.wikia.tibia.objects.TibiaWikiBot;
-import com.wikia.tibia.repositories.WikiArticleRepository;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+
+
 public class FixCreaturesTest {
 
     private FixCreatures target;
-    private WikiArticleRepository repository;
 
     @Before
     public void setup() {
-        TibiaWikiBot tibiaWikiBot = new TibiaWikiBot();
-        tibiaWikiBot.login();
-        target = new FixCreatures(tibiaWikiBot);
-        this.repository = new WikiArticleRepository(tibiaWikiBot);
+        target = new FixCreatures();
     }
 
     @Ignore
@@ -26,5 +24,6 @@ public class FixCreaturesTest {
         // when
         target.checkCreatures();
         // then
+        assertThat("Done", is("Done"));
     }
 }
